Create class Enemy. Add file spriteSheetEnemyZero40x40.png. Update classes Leve1State and TileMap

+++ SvartSlott/src/com/marc/tilemap/TileMap.java
@@ -117,14 +117,14 @@ public class TileMap
         return tileSize;
     }
 
-    public int getAxisX()
+    public double getAxisX()
     {
-        return (int)axisX;
+        return axisX;
     }
 
-    public int getAxisY()
+    public double getAxisY()
     {
-        return (int)axisY;
+        return axisY;
     }
 
     public int getMapWidth()
